FOG: Very primitive idea implemented
? Realizing I need to make visible not the squares of the possible legal 
moves but the squares but also the immediately ajacent square in front, 
and for pawns also the idea of the square two squares ahead (on first 
move) even if it's blocked by a piece? Or can a pawn always only see one 
square ahead? That would make no sense at all. I think this is only a 
problem for pawns? What... hmm. This is complicated.