FOG: "Better" version that roughly works
? Currently we have the version where pieces see every ajacent square + 
their possible move squares. It turns out to reveal a lot of the board 
pretty quickly, which I don't know what I love, but I'm feeling quite 
unclear on what a piece can "see" in the first place really, what that 
means in the context of chess where it's normally a game of total 
information.