FOG: Got a more sophisticated, gradiated fog going
? Now every piece that can see a square (adjacent + moves) makes its fog 
a little less opaque, leading to a nicer looking board and a much more 
mysterious experience. There's definitely potential strategy here. Found 
an early move that led to a queen being captured by a knight it could 
see which is kind of great... knights are both vulnerable and powerful 
based on line of sight.