LEWITT: Added most basic idea of saving out the image on game over
? It works perfectly well but it has a transparent background because 
I'm using a transparent image, and I'd prefer that it end up with white 
or even a wall-textured image? To do so I guess I need an image with the 
background and then to copy all non-clear pixels from canvas across and 
then save that image out (maybe by rendering it back onto the 
canvas...?). Doable.