DRAUGHTS: Multiple jumps now works
? It's bloody satisfying to watch multiple captures working properly 
(for linear pieces). Once the queen gets going you are in deep shit 
potentially, which looks quite incredible.

? I'm unclear on how to deal with check and checkmate. Technically if 
captures only occur when adjacent, you should only be in check when the 
threatening piece is next to the king with a space to jump to? But then 
a king in the corner is not capturable... which is kind of funny? You're 
in checkmate when you can't move your king to a square where there isn't 
a piece that can jump it. But it may then be possible to just jump in 
and out of a corner?

? How much am I concerned with the idea of stalemate and endgames versus 
the basic satisfaction of checkers capturing on a chess board? I have to 
be a little bit, since I want a complete game, but perhaps if the 
endgame sucks that's just the way it is? If you captured all your 
opponents pieces then would you be able to force a checkmate? Possibly. 
And maybe king in the corner is just a crucial piece of strategy that 
you need to include when you play?

? It strikes me there's a lot of extra potential complexity involved in 
calculating check here because it has literally different rules, which 
may be a bit ugh. But I'll try.