Starte info panel work
? This kind of ui programming isn't exactly my favourite element of 
making a game, but it's important in this instance to offer a bit of 
information to give players the gist of what's going on.

? I think it's best to only display the info option when they start 
playing as that feels kind of fun. But there will no doubt be some 
issues no matter what I do. The difference between touch and mouse is a 
key issue (no hover most obviously).

? I think I'll conquer this, it just might take longer than I want.