MUSICAL: Added visual note indicators
? There's a current file indicator as well as a specific highlight of 
the pieces being played. It should make a decent amount of sense.

? It's also true that now the note indicators basically _are_ the turn 
indicator, so may not be necessary to have a separate turn bar in this 
game? If I remove it the key problem I foresee is that it's there as an 
indicator of allowable input, but I doubt anyone is playing so fast they 
would notice?

? So in the case let me add that in too. And it's nice!