CHECK-RS: Game ends on capture of a king
? In the end I think this is the best decision for decent play. Couple 
of nice things: a) it lets you force moves by the king so that it jumps 
itself into a capture (possibly via multiple jumps) which is kind of 
fun, and b) it might be a way out of a stalemate with the king hugging 
the corner, because you lure the king out with one piece and then jump 
with another.

? I almost wonder if I can say I know stalemate is just two kings and 
any other set of pieces is winnable?